Move LogEnvironmentInfo to StartupHelpers
